+++ lib/riot_api/global/account.rb
@@ -6,24 +6,13 @@ require_relative 'account/by_riot_id'
 module RiotApi
   module Global
     module Account
-      include ResponseAttributes
+      include RiotApi::Global::ResponseAttributes
 
       REGIONS = %w[europe americas asia esports].freeze
 
       def call
         check_region
-
-        response = send_request
-
-        wrap_response(response)
-      end
-
-      private
-
-      def check_region
-        return if REGIONS.include?(region)
-
-        raise InvalidRegionError, 'Invalid region: must be europe, americas, asia or esports'
+        wrap_response(send_request)
       end
     end
   end